python/requirements.txt
django==1.10
djangorestframework==3.6.4
django-cors-headers==2.1.0
django-extensions==1.9.0
mysqlclient==1.3.12
djangorestframework-jwt==1.11
Werkzeug==0.12.2
numpy==1.13.1
pandas==0.20.3
matplotlib==2.0.2